+++ CHANGELOG.md
@@ -34,4 +34,6 @@
 
 2019-12-22 Eric Lau
 
+    * clean up import mess
     * add back button to auth pages
+